6838583 samdb check fails on inodes with even gen
Modified how to determine in-use inodes by checking for an empty ip->di->mode or an
extention inode.  Previously it was assumed that an odd gen number meant the inode
was in-use. The changes for this mod are in src/utility/samdb/check.c in inode_callback().
Also a new header file was included ino_ext.h to check S_ISEXT.

SourceCset: 6bea64f3e55c