6826981 schproj ignores path access permissions
Fix issue with schproj where it allows one to schproj files that are
hidden below directories for which the user has no search permission.
This is caused by schproj running suid root.

The check_access() routine is too complicated.  It should be
replaced with a call to access(2), but the Solaris access() call
is broken the same way (see CR 1187647).  This CR is fixed / integrated
into snv_116.  When we are sure we are not going to run on any level
lower than 116 (will be some time), should either change check_access()
to use access(2), or pull this mod entirely (may be taken care of by
the fix for 1187647).

SourceCset: 910394d85f8c