6839852 samdb check samqfs1 -s dumps core file on newly sammkfs'd filesystem
An assumption was made that allocation would always occur, which could result in
accessing a bad pointer.  Modified to check the size of the list before accessing it.
Changes are to src/utility/samdb/check.c check_directory() and a memory initialization
at the beginning of samdb_check().

SourceCset: 2db6cc23173a