6825050 Kernel cmn_err indicates wrong method for upgrading file system from 2 to 2A.
When using a new 5.0 feature, if the file system is V2, it issues a
kernel cmn_err of the form: "Can't add/remove/release.  Upgrade file
system with samadm add-features first.".  The method of upgrading was
changed to "samfsck -u".  This change updates the text of the to
indicate the proper way.

SourceCset: bf86255ab4d2