6850566 host on/off with Sun Cluster, FS failover to host marked off
SAM-QFS SC agent was not noticing that a node was marked "off"
in the hosts file, and allowing a failover to that node.  Technically,
this is unsupported because "off" hosts are supposed to be down
already (see man page for samsharefs), but we could do better here
and check in the agent for off hosts.

SourceCset: f9c7788336ab