6871959 fcntl(F_SETLKW) and lockf() hang when they should fail with errno EDEADLK
First try acquiring the lock as today, by converting to non-blocking requests.
If we cannot acquire the lock and it is a blocking request, then try to start
an asynchronous task to wait for the lock and immediately return to the client
with SR_WAIT_LEASE as before.

We also need to keep track of the tasks that have been started to cancel them
during failover and unmount.

SourceCset: 073d844e5150