6842301 Non-zero data beyond EOF with small 4K DAU can cause corruption if pages invalidated
For md devices on SPARC, (4k DAUs in a 8K page), zero the
uninitialized portion of the upper 4K block in the 8k page
of the last page in this loop. This may be done more than
once for different user requests, but is needed here to
avoid a page deadlock.

SourceCset: 1363bd75f398