6891698 Non-zero data beyond EOF with small 4K DAU can cause corruption if pages invalidated
The code added by 6532843 is not needed. The page is zeroed when appending
with SAM_FORCEFAULT set and writing in the last 4K of an 8K page.
This code is not needed because it was fixed by the code in sam_allocate_block.
This bug removes the code in sam_write_io which was added by bug 6532843.

SourceCset: 20a4d7056150