samfs: number of blocks is stored in inode as int32
on large files, greater than 8TB this value can be negative. In upper OS
filesystem layer this value is assigned to uin32t and turn into a value
larger than MAXLONGINT.
To prevent this, this  change will first cast the int32 to uint32 before it is
assinged to the uint64.